+++ validator-core/src/main/java/no/difi/vefa/validator/ValidatorEngine.java
@@ -1,6 +1,7 @@
 package no.difi.vefa.validator;
 
 import no.difi.vefa.validator.api.SourceInstance;
+import no.difi.vefa.validator.api.ValidatorException;
 import no.difi.xsd.vefa.validator._1.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;