+++ validator-core/src/main/java/no/difi/vefa/validator/api/Checker.java
@@ -3,7 +3,6 @@ package no.difi.vefa.validator.api;
 import no.difi.vefa.validator.Configuration;
 import no.difi.vefa.validator.Document;
 import no.difi.vefa.validator.Section;
-import no.difi.vefa.validator.ValidatorException;
 
 import java.nio.file.Path;