+++ validator-core/src/main/java/no/difi/vefa/validator/api/Presenter.java
@@ -1,7 +1,6 @@
 package no.difi.vefa.validator.api;
 
 import no.difi.vefa.validator.Document;
-import no.difi.vefa.validator.ValidatorException;
 import no.difi.xsd.vefa.validator._1.StylesheetType;
 
 import java.io.OutputStream;