+++ validator-core/src/main/java/no/difi/vefa/validator/api/ReloadableContext.java
@@ -1,0 +1,7 @@
+package no.difi.vefa.validator.api;
+
+public interface ReloadableContext {
+
+    void load() throws ValidatorException;
+
+}