+++ validator-core/src/main/java/no/difi/vefa/validator/api/Source.java
@@ -1,7 +1,5 @@
 package no.difi.vefa.validator.api;
 
-import no.difi.vefa.validator.ValidatorException;
-
 /**
  * Source for validation artifacts.
  */
@@ -10,6 +8,7 @@ public interface Source {
     /**
      * Instance of source with validation artifacts ready for use.
      *
+     * @throws ValidatorException
      * @return Instance containing validation artifacts.
      */
     SourceInstance createInstance() throws ValidatorException;