+++ validator-core/src/main/java/no/difi/vefa/validator/api/ValidatorException.java
@@ -1,4 +1,4 @@
-package no.difi.vefa.validator;
+package no.difi.vefa.validator.api;
 
 /**
  * Exception specific to validator.