+++ validator-core/src/main/java/no/difi/vefa/validator/checker/XsdChecker.java
@@ -3,7 +3,7 @@ package no.difi.vefa.validator.checker;
 import no.difi.vefa.validator.Configuration;
 import no.difi.vefa.validator.Document;
 import no.difi.vefa.validator.Section;
-import no.difi.vefa.validator.ValidatorException;
+import no.difi.vefa.validator.api.ValidatorException;
 import no.difi.vefa.validator.api.Checker;
 import no.difi.vefa.validator.api.CheckerInfo;
 import no.difi.vefa.validator.util.PathLSResolveResource;