+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Functionality for configuring validator.
 * Supporting parameters for stylesheets.
 * Better defaults in the sample application (validator-web).
+* More Javadoc in code.
+* ValidatorException is moved from no.difi.vefa.validator to no.difi.vefa.validator.api.
+* Cleaner use of exceptions in exposed classes.
 
 ## 2.0.0-RC1